breaking change: Add `BaseController`, use a `BaseController.Context` type to pass parameters to controllers instead of setting members before invoking their methods, add `ResourceRegistry.store` property to give access to the underlying store.
