+++ examples/local_file/main.py
@@ -75,4 +75,4 @@ ctx.resource_types.register(LocalFile)
 ctx.controllers.register(LocalFileController())
 ctx.load_manifest(Path(__file__).parent / "manifest.yaml")
 # ctx.resources.delete(LocalFile.uri("default", "local-file"), do_raise=False)
-ctx.controllers.reconcile()
+ctx.reconcile()
