+++ src/equilibrium/NotSet.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class NotSet(Enum):
+    Value = 1
