+++ src/equilibrium/__init__.py
@@ -3,23 +3,23 @@ Equilibrium is a Python framework for handling Kubernetes-like resources and imp
 """
 
 from equilibrium.AdmissionController import AdmissionController
+from equilibrium.BaseController import BaseController
+from equilibrium.ControllerRegistry import ControllerRegistry
 from equilibrium.CrudResourceController import CrudResourceController
 from equilibrium.JsonResourceStore import JsonResourceStore
 from equilibrium.Namespace import Namespace
 from equilibrium.Resource import Resource
-from equilibrium.ResourceContext import (
-    ControllerRegistry,
-    ResourceContext,
-    ResourceRegistry,
-    ResourceTypeRegistry,
-    ServiceRegistry,
-)
+from equilibrium.ResourceContext import ResourceContext
 from equilibrium.ResourceController import ResourceController
+from equilibrium.ResourceRegistry import ResourceRegistry
 from equilibrium.ResourceStore import ResourceStore
+from equilibrium.ResourceTypeRegistry import ResourceTypeRegistry
 from equilibrium.Service import Service
+from equilibrium.ServiceRegistry import ServiceRegistry
 
 __all__ = [
     "AdmissionController",
+    "BaseController",
     "ControllerRegistry",
     "CrudResourceController",
     "JsonResourceStore",
